+++ ui/src/components/dialog/SettingsDialog.vue
@@ -2,6 +2,20 @@
   <Dialog :title="t('settingsTitle')" :show="show" @close="close" :width="`700px`" :goBack="close">
     <template v-slot:content>
       <v-card-text class="pt-4">
+        <h3>{{ t('keyBackupTitle') }}</h3>
+        <p>
+          {{ t('loseAccess') }}.
+          {{ t('backUpKey') }}
+        </p>
+        <v-col style="margin-bottom:20px" align="left" class="py-0">
+          <v-btn
+            class="white--text"
+            color="black"
+            @click.prevent="downloadKeys"
+          >
+            {{ t('downloadKey') }}
+          </v-btn>
+      </v-col>
         <h3>{{ t('languageTitle') }}</h3>
         <p>
           {{ t('chooseLanguage') }}
@@ -24,6 +38,12 @@
 import Dialog from '@/components/dialog/Dialog.vue'
 import LocalePicker from '@/components/LocalePicker'
 
+import { mapGetters, mapActions, createNamespacedHelpers } from 'vuex'
+import { downloadKeys } from '@/lib/key-backup.js'
+
+const { mapActions: mapSettingsActions } = createNamespacedHelpers('settings')
+const { mapActions: mapTribeActions } = createNamespacedHelpers('tribe')
+
 export default {
   props: {
     show: { type: Boolean, required: true }
@@ -33,7 +53,27 @@ export default {
     Dialog,
     LocalePicker
   },
+  computed: {
+    ...mapGetters(['whoami'])
+  },
   methods: {
+    ...mapSettingsActions(['updateSettings', 'getLatestSeq']),
+    ...mapTribeActions(['getTribeIds']),
+    ...mapActions(['setWhoami']),
+    async downloadKeys () {
+      const groupIds = await this.getTribeIds()
+      const latestMssgSeq = await this.getLatestSeq()
+
+      downloadKeys(groupIds, latestMssgSeq)
+
+      const input = {
+        id: this.whoami.personal.settings.id,
+        keyBackedUp: true
+      }
+      await this.updateSettings(input)
+
+      await this.setWhoami()
+    },
     close () {
       this.$emit('close')
     },