+++ ui/src/components/profile/ProfileForm.vue
@@ -395,6 +395,7 @@ import DateIntervalPicker from '@/components/DateIntervalPicker.vue'
 
 import { GENDERS, RELATIONSHIPS } from '@/lib/constants'
 import { getDisplayName } from '@/lib/person-helpers.js'
+import { downloadKeys } from '@/lib/key-backup.js'
 
 import isEmpty from 'lodash.isempty'
 
@@ -511,6 +512,7 @@ export default {
   },
   methods: {
     getDisplayName,
+    downloadKeys,
     updateSelectedGender (genderClicked) {
       // reset images to outlined
       this.$refs.taneImg.src = require('@/assets/tane-outlined.svg')